MAGECLOUD-1327: Move cache clean operation in deploy phase to the very end
